feat: share app theme and feiwen titlebar
