+++ app/ai-chat/Cargo.toml
@@ -43,9 +43,12 @@ gpui.workspace = true
 gpui_platform.workspace = true
 gpui-component.workspace = true
 gpui-component-assets = { workspace = true }
+app-theme = { workspace = true, default-features = false, features = [
+  "system-accent",
+  "theme-picker",
+] }
 window-ext = { workspace = true }
 platform-ext = { workspace = true }
-material-color-utils = { version = "0.1.3", default-features = false }
 rust-embed = { version = "8.11.0", features = ["interpolate-folder-path"] }
 anyhow = "1.0.102"
 tray-icon = { version = "0.24.0", default-features = false, features = [
