+++ app/feiwen/Cargo.toml
@@ -10,6 +10,7 @@ gpui_platform.workspace = true
 gpui-component.workspace = true
 gpui-component-assets.workspace = true
 app-assets.workspace = true
+app-theme = { workspace = true, default-features = false, features = ["system-accent"] }
 fluent-bundle = "0.16.0"
 unic-langid = "0.9.6"
 sys-locale = "0.3.2"
