+++ crates/app-theme/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "app-theme"
+version = "0.1.0"
+edition = "2024"
+
+[features]
+default = ["system-accent", "theme-picker"]
+system-accent = ["dep:platform-ext"]
+theme-picker = []
+
+[dependencies]
+gpui.workspace = true
+gpui-component.workspace = true
+material-color-utils = { version = "0.1.3", default-features = false }
+platform-ext = { workspace = true, optional = true }
+serde_json = "1.0.149"
+smol = "2.0.2"
